requirements.txt
pillow
requests
tweepy
python-dotenv
google-generativeai

